refactor(interpreter): split the lisp.ts monolith into modules
Break the ~2050-line lisp.ts into focused modules (sexpr, printer,
exceptions, schemas, func, compile, interp, builtins, reader, repl, io,
main) with lisp.ts kept as the public re-export barrel, so no consumer
import changes. The prelude moves out of an escaped template string into
real Lisp source (prelude.ptc) loaded at import time.

Along the way: the built-in registry moves out of the Interp constructor
into grouped registrars; magic numbers become named constants; the
TS-level mapcar is renamed mapList to stop shadowing the prelude's Lisp
mapcar; truncate throws a proper EvalException instead of a string; and
divide's Number coercion is documented as intentional (Lisp / is true
division).
